Makes new Vue templates use faaslang
