+++ templates/vue/_files/app/static/index.html
@@ -11,10 +11,10 @@
       newpath = newpath.endsWith('/') ? newpath : newpath + '/';
       (newpath !== pathname) && (window.location.pathname = newpath);
     </script>
-    <script src="?env"></script>
+    <script src="static/?env=true"></script>
   </head>
   <body>
     <div id="app"></div>
-    <script src="static/?path=main.js"></script>
+    <script src="static/main.js"></script>
   </body>
 </html>
