--- templates/vue/_files/f/main/function.json
@@ -1,11 +1,0 @@
-{
-  "name": "main",
-  "description": "Retrieves your index.html or env variables",
-  "args": [],
-  "kwargs": {},
-  "http": {
-    "headers": {
-      "Content-Type": "text/html"
-    }
-  }
-}
