--- templates/vue/_files/f/static/function.json
@@ -1,13 +1,0 @@
-{
-  "name": "static",
-  "description": "Retrieves files from your static directory.",
-  "args": [],
-  "kwargs": {
-    "path": "The path to the retrieved file, relative to the static directory."
-  },
-  "http": {
-    "headers": {
-      "Content-Type": "text/html"
-    }
-  }
-}
